requirements.txt
# requirements.txt (Phần được cập nhật)

# --- Data Core & Handling ---
pandas
numpy

# --- Pillar 1: Contract Risk Model ---
networkx
slither-analyzer  # 
etherscan-api  # 

# --- Pillar 2: Gas Cost Model ---
statsmodels
tensorflow
keras
scikit-learn

# --- Pillar 3: User Behavior Model ---
# (networkx, scikit-learn đã có)

# --- Data Acquisition & Utilities ---
web3
requests
matplotlib
seaborn
python-dotenv     # <--- THÊM VÀO: Cần cho config
google-cloud-bigquery
google-auth
db-dtypes